==> Assignments/A1/step2-scrape_presidential_party_and_term.ipynb <==
import pandas as pd
from selenium import webdriver
from selenium.webdriver.firefox.service import Service
from webdriver_manager.firefox import GeckoDriverManager
import re
import dateparser
import datetime

# here we scrape information on president's term and party
# 
# NOTE: Britiania seems to be attempting to block web scrapers. When this happened, a regular requests
# approach will fail. To bypass this, you will need to use  selenium. 
# The following code should work if site is blocking scraper:

# Start a driver session....

    # if you have selenium 3 installed, use one of these:
#driver = webdriver.Firefox(executable_path=GeckoDriverManager().install()) # this will work on Windows and Mac, and should work on Linux when run the first time
#driver = webdriver.Firefox(executable_path=<insert path to manual downloaded geckodriver>)
#driver = webdriver.Firefox() # use if geckodriver is in your PATH environmnet variable (which includes the same folder as your notebook)

    # if you hve selenium 4 installed, use one of these:
#driver = webdriver.Firefox(service=Service(GeckoDriverManager().install())) # this will work on Windows and Mac, and should work on Linux when run the first time
#driver = webdriver.Firefox() # use if geckodriver is in your PATH environmnet variable (which includes the same folder as your notebook)
driver = webdriver.Firefox(service=Service("../../geckodriver")) # add your own path to geckodriver


driver.get("https://www.britannica.com/topic/Presidents-of-the-United-States-1846696")
driver.implicitly_wait(10)
page_source = driver.page_source
driver.close() 

# pandas read html will parse the contents of the table in the downloaded webpage
presidents = pd.read_html(page_source)[0]
presidents

# note that the last two rows contains non-presidential information
# let's remove these last two rows...
presidents = presidents.drop([int(len(presidents)-1), int(len(presidents)-2)])
presidents

In the following cell, we create two new columns - from and to. We parse the term column for the information we need, and format appropriately.

presidents['term'] = presidents.apply(lambda row: re.sub(r"[^\d–]", "", row['term']), axis=1) # replace any non-digit before dash with blank

#re.sub('[^0-9–]', '',  "1897-1901*")

presidents


datetime.datetime.now().year

# note: this cell has been updated since the video was recorded. There was an issue with the previous code
# where I used dash rather than en-dash character (they basically look the same, but en-dash is wider)
# Also, I updated the logic to use all apply methods. I think this logic is a bit easier to follow.

# # remove all characters except digits and dashes
presidents['term'] = presidents.apply(lambda row: re.sub(r"[^\d–]", "", row['term']), axis=1) # replace any non-digit before dash with blank

# first, split the string in the term column using dash as delimiter - store this in new column called 'from'
presidents['from'] = presidents['term'].apply(lambda x: x.split("–")[0]) 

# if there is no end year, then the end year is the same as the start year, otherwise get the second element of the split string, this is the end year of the term
presidents['to'] = presidents['term'].apply(lambda x: x.split("–")[0] if len(x.split("–"))==1 else x.split("–")[1]) 

# if the end year is only two digits, then add the first two digits of the start year to the end year
presidents['to'] = presidents.apply(lambda row:  row['from'][:2]+row['to'] if len(row['to'])==2 else row['to'], axis=1) 

presidents

# this cell is new since the video.
# As I mentioned in the video, you can drop the unused/unecessary columns
presidents = presidents.drop(columns=['Unnamed: 0', 'no.','term']) # drop the term column
presidents

# this cell is also new since the video. This code adds a to date that is missing for the last president in the list.
# it does this by adding 4 years to the from date
presidents.loc[presidents.index[-1], 'to'] = int(presidents.loc[presidents.index[-1], 'from'])+4 # the last president's term is 4 years, so we can calculate the end year

presidents.to_csv("./data/presidential_party_and_term.csv", index=False)
